Implemented Flask JWT Extended

+++ resources/item.py
@@ -1,11 +1,9 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt, get_jwt_identity
 from models.item import ItemModel
 
 
 class Item(Resource):
-    TABLE_NAME = 'items'
-
     parser = reqparse.RequestParser()
     parser.add_argument('price',
         type=float,
@@ -19,12 +17,14 @@ class Item(Resource):
                         )
 
     @jwt_required()
+    # Headers:      Authorization : Bearer xxxxx
     def get(self, name):
         item = ItemModel.find_by_name(name)
         if item:
             return item.json()
         return {'message': 'Item not found'}, 404
 
+    @jwt_required(fresh=True)
     def post(self, name):
         if ItemModel.find_by_name(name):
             return {'message': "An item with name '{}' already exists.".format(name)}
@@ -42,6 +42,9 @@ class Item(Resource):
 
     @jwt_required()
     def delete(self, name):
+        claims = get_jwt()
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required.'}, 401
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -64,7 +67,17 @@ class Item(Resource):
 
 
 class ItemList(Resource):
-    TABLE_NAME = 'items'
-
+    @jwt_required(optional=True)
     def get(self):
-        return {'items': [item.json() for item in ItemModel.find_all()]}
+        """
+        Here we get the JWT identity, and then if the user is logged in (we were able to get an identity)
+        we return the entire item list. Otherwise we just return the item names.
+        """
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.find_all()]
+        if user_id:
+            return {'items': items}, 200
+        return {
+            'items': [item['name'] for item in items],
+            'message': 'More data available if you log in.'
+        }, 200
